Pulled Option & OptionMap classes into engine.py for common use
